BACKEND/requirements.txt
Flask==3.1.2
Flask-MySQLdb==2.0.0
mysqlclient==2.2.7
PyJWT==2.10.1
Flask-Cors==6.0.1
Flask-Limiter==4.0.0
python-dotenv==1.2.1
Werkzeug==3.1.3
